requirements.txt
flask==3.0.3
selenium==4.20.0
pytest==8.2.2
pytest-html==4.1.1
webdriver-manager==4.0.2
PyYAML==6.0.1
